dspy-fact-checker/pyproject.toml
[project]
name = "dspy-fact-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy-ai>=3.0.3",
    "ipykernel>=7.0.1",
    "wikipedia>=1.4.0",
]
